+++ changelog.md
@@ -1,2 +1,6 @@
+## Additions
+- Health bars on the HUD now display the status effects of the target.
+- Added a config option to toggle status effects on the health bar HUD.
+
 ## Bugfixes
 - Fixed an issue that caused all nameplates to be permanently visible.